refactor(system): overhaul — cleanup broken APIs, add ProcessMonitor + SystemLoad
Phase 1 — Cleanup:
- Remove broken memory::management submodule (allocate/free/protect/lock/unlock/query)
- Remove deps: region, once_cell, tokio; features: component, metrics, async, memory
- Remove broken network stubs: connections, is_online, config, detect_gateway
- Remove dangerous process APIs: kill, set_priority; remove always-empty cmd/environ fields
- Delete dead SYSTEM_INFO_CACHE + SystemInfo::refresh()
- Fix process::stats() stale read (now refreshes before reading)
- Fix NUMA inconsistency (info::detect_numa_nodes reads /sys/ on Linux)
- Replace once_cell::Lazy with std::sync::LazyLock in network.rs

Phase 2 — Test hardening:
- Add 20 new tests: pressure boundary values, structural invariants, error paths
- Coverage for cpu, memory, disk, network, process, info modules (38 total)

Phase 3 — ProcessMonitor for sandbox per-PID tracking:
- ProcessMonitor::new(pid), sample() → ProcessSample, peak_memory(), elapsed()

Phase 4 — SystemLoad for adaptive worker scaling:
- system_load() → SystemLoad with can_accept_work() and headroom()

Phase 5 — Delete outdated crates/system/design/ folder (7 files)

+++ crates/system/src/utils.rs
@@ -84,9 +84,11 @@ pub fn format_duration(duration: std::time::Duration) -> String {
     if nanos < 1_000 {
         format!("{nanos}ns")
     } else if nanos < 1_000_000 {
-        format!("{:.2}µs", nanos as f64 / 1_000.0)
+        // Cast through u64 first — nanos < 1_000_000 fits in u64,
+        // avoids __floattidf call for u128→f64 conversion.
+        format!("{:.2}µs", nanos as u64 as f64 / 1_000.0)
     } else if nanos < 1_000_000_000 {
-        format!("{:.2}ms", nanos as f64 / 1_000_000.0)
+        format!("{:.2}ms", nanos as u64 as f64 / 1_000_000.0)
     } else {
         let secs = duration.as_secs();
         if secs < 60 {
